# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CareLens API",
    "description": "Clinical decision-support API for plain radiograph bone tumor detection and Grad-CAM interpretability.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/cases": {
      "post": {
        "tags": [
          "Clinical Cases"
        ],
        "summary": "Submit Case",
        "description": "Submit plain radiograph image, execute ResNet-50 inference + Grad-CAM,\nand persist new clinical case record.",
        "operationId": "submit_case_cases_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_submit_case_cases_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Clinical Cases"
        ],
        "summary": "List Cases",
        "description": "List cases:\n- Patients: view their own submitted scans\n- Doctors: view triage queue and reviewed cases with optional status filter",
        "operationId": "list_cases_cases_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter: 'pending', 'reviewed', or specific status",
              "title": "Status"
            },
            "description": "Filter: 'pending', 'reviewed', or specific status"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedCasesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cases/stats": {
      "get": {
        "tags": [
          "Clinical Cases"
        ],
        "summary": "Get Stats",
        "description": "Fetch aggregated clinical dashboard counters.",
        "operationId": "get_stats_cases_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardStats"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/cases/{case_id}": {
      "get": {
        "tags": [
          "Clinical Cases"
        ],
        "summary": "Get Case Detail",
        "description": "Fetch complete case details (enforces per-case ownership / doctor role).",
        "operationId": "get_case_detail_cases__case_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cases/{case_id}/review": {
      "post": {
        "tags": [
          "Clinical Cases"
        ],
        "summary": "Submit Doctor Review",
        "description": "Submit formal clinical review (Doctor role required).",
        "operationId": "submit_doctor_review_cases__case_id__review_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cases/{case_id}/report.pdf": {
      "get": {
        "tags": [
          "Clinical Cases"
        ],
        "summary": "Download Case Report",
        "description": "Download comprehensive clinical PDF report for a case.",
        "operationId": "download_case_report_cases__case_id__report_pdf_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/cases": {
      "post": {
        "tags": [
          "Clinical Cases"
        ],
        "summary": "Submit Case",
        "description": "Submit plain radiograph image, execute ResNet-50 inference + Grad-CAM,\nand persist new clinical case record.",
        "operationId": "submit_case_api_cases_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_submit_case_api_cases_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Clinical Cases"
        ],
        "summary": "List Cases",
        "description": "List cases:\n- Patients: view their own submitted scans\n- Doctors: view triage queue and reviewed cases with optional status filter",
        "operationId": "list_cases_api_cases_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter: 'pending', 'reviewed', or specific status",
              "title": "Status"
            },
            "description": "Filter: 'pending', 'reviewed', or specific status"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedCasesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cases/stats": {
      "get": {
        "tags": [
          "Clinical Cases"
        ],
        "summary": "Get Stats",
        "description": "Fetch aggregated clinical dashboard counters.",
        "operationId": "get_stats_api_cases_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardStats"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/cases/{case_id}": {
      "get": {
        "tags": [
          "Clinical Cases"
        ],
        "summary": "Get Case Detail",
        "description": "Fetch complete case details (enforces per-case ownership / doctor role).",
        "operationId": "get_case_detail_api_cases__case_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cases/{case_id}/review": {
      "post": {
        "tags": [
          "Clinical Cases"
        ],
        "summary": "Submit Doctor Review",
        "description": "Submit formal clinical review (Doctor role required).",
        "operationId": "submit_doctor_review_api_cases__case_id__review_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cases/{case_id}/report.pdf": {
      "get": {
        "tags": [
          "Clinical Cases"
        ],
        "summary": "Download Case Report",
        "description": "Download comprehensive clinical PDF report for a case.",
        "operationId": "download_case_report_api_cases__case_id__report_pdf_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_submit_case_api_cases_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_submit_case_api_cases_post"
      },
      "Body_submit_case_cases_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_submit_case_cases_post"
      },
      "CaseResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "patient_id": {
            "type": "integer",
            "title": "Patient Id"
          },
          "patient_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Name"
          },
          "image_url": {
            "type": "string",
            "title": "Image Url"
          },
          "gradcam_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gradcam Url"
          },
          "model_prediction": {
            "type": "string",
            "title": "Model Prediction"
          },
          "cancer_probability": {
            "type": "number",
            "title": "Cancer Probability"
          },
          "risk_band": {
            "type": "string",
            "title": "Risk Band"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "doctor_verdict": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doctor Verdict"
          },
          "doctor_explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doctor Explanation"
          },
          "recommendation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "patient_id",
          "image_url",
          "model_prediction",
          "cancer_probability",
          "risk_band",
          "status",
          "created_at"
        ],
        "title": "CaseResponse"
      },
      "CaseReviewRequest": {
        "properties": {
          "doctor_verdict": {
            "type": "string",
            "minLength": 2,
            "title": "Doctor Verdict",
            "description": "Physician final verdict (e.g. Malignant, Benign, Indeterminate)"
          },
          "doctor_explanation": {
            "type": "string",
            "minLength": 5,
            "title": "Doctor Explanation",
            "description": "Clinical findings and radiologic interpretation"
          },
          "recommendation": {
            "type": "string",
            "minLength": 5,
            "title": "Recommendation",
            "description": "Next actionable clinical recommendation (e.g. Biopsy, CT, Follow-up)"
          }
        },
        "type": "object",
        "required": [
          "doctor_verdict",
          "doctor_explanation",
          "recommendation"
        ],
        "title": "CaseReviewRequest"
      },
      "DashboardStats": {
        "properties": {
          "pending_count": {
            "type": "integer",
            "title": "Pending Count"
          },
          "reviewed_count": {
            "type": "integer",
            "title": "Reviewed Count"
          },
          "high_risk_count": {
            "type": "integer",
            "title": "High Risk Count"
          },
          "total_patients": {
            "type": "integer",
            "title": "Total Patients"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "device": {
            "type": "string",
            "title": "Device"
          }
        },
        "type": "object",
        "required": [
          "pending_count",
          "reviewed_count",
          "high_risk_count",
          "total_patients",
          "model_version",
          "device"
        ],
        "title": "DashboardStats"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedCasesResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CaseResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "limit",
          "total_pages"
        ],
        "title": "PaginatedCasesResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}